export addRoundedRectPath
closes #8907
