+++ packages/tldraw/api-report.api.md
@@ -186,6 +186,9 @@ export interface ActionsProviderProps {
     overrides?(editor: Editor, actions: TLUiActionsContextType, helpers: TLUiOverrideHelpers): TLUiActionsContextType;
 }
 
+// @public
+export function addRoundedRectPath(path: Path2D, bounds: Box, radius: number, counterClockwise?: boolean): void;
+
 // @public (undocumented)
 export type AlertSeverity = 'error' | 'info' | 'success' | 'warning';
